Create README.md
These files are Cullan Crothers modified versions of examples from Imar Spaanjaars book on ASP.NET 4.5.1. They are intended for viewing only. The repository was created at the request of a job recruitment company that wants to see some of the work that I have done from the book examples.